requirements.txt
fastapi
uvicorn
pytesseract
pillow
reportlab
python-multipart
